Section_06/22_Email Scheduler/pyproject.toml
[project]
name = "22-email-scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "schedule>=1.2.2",
]
